Add ChatProcessorService to manage conversation processing logic
This update introduces the ChatProcessorService class, which handles the processing of chat prompts and interactions with the AI client. By modularizing this functionality, it improves code organization and maintainability, while also allowing for more flexible conversation management capabilities. Additional dependencies for this service have also been included in the main AIA module.
